Move api interfaces to a subfolder
