+++ lib/app.js
@@ -3,9 +3,10 @@
 /** Dependecies **/
 var express = require('express');
 var routes = require('./routes/index');
-var sabroutes = require('./routes/sabnzbd');
-var sickbeardRoutes = require('./routes/sickbeard');
 var sitesRoutes = require('./routes/sites');
+// API interfaces
+var sabRoutes = require('./routes/api/sabnzbd');
+var sickbeardRoutes = require('./routes/api/sickbeard');
 
 // Config
 var config = {
@@ -22,7 +23,7 @@ app.engine('jsx', require('express-react-views').createEngine());
 app.set('views', __dirname + '/views');
 app.use(express.static(__dirname + '/public'));
 
-app.use('/sabnzbd', sabroutes);
+app.use('/sabnzbd', sabRoutes);
 app.use('/sickbeard', sickbeardRoutes);
 app.use('/sites', sitesRoutes);
 app.use('/', routes);
